+++ bubolo/src/bubolo/world/entity/concrete/Base.java
@@ -12,7 +12,7 @@ import bubolo.world.entity.StationaryElement;
 /**
  * Bases allow Tanks to heal and recover their mines, and capturing them is the primary
  * goal of the game.
- * 
+ *
  * @author BU CS673 - Clone Productions
  */
 public class Base extends StationaryElement implements Ownable, Damageable
@@ -38,31 +38,31 @@ public class Base extends StationaryElement implements Ownable, Damageable
 
 	/**
 	 * Boolean representing whether this Base is currently charging a Tank.
-	 * 
+	 *
 	 */
 	private boolean isCharging = false;
 
 	/**
 	 * The current number of hit points of the base.
 	 */
-	private int hitPoints;
-	
+	private float hitPoints;
+
 	private static final int MAX_HIT_POINTS = 100;
-	
+
 	private static final int HIT_POINTS_REPLENISH_RATE = 5;
-	
+
 	private int mineCount;
-	
+
 	private static final int MAX_MINE_COUNT = 10;
-	
+
 	private static final int MINE_REPLENISH_RATE = 1;
-	
+
 	private int ammoCount;
 
 	private static final int MAX_AMMO_COUNT = 100;
-	
+
 	private static final int AMMO_REPLENISH_RATE = 5;
-	
+
 	/**
 	 * Construct a new Base with a random UUID.
 	 */
@@ -73,7 +73,7 @@ public class Base extends StationaryElement implements Ownable, Damageable
 
 	/**
 	 * Construct a new Base with the specified UUID.
-	 * 
+	 *
 	 * @param id
 	 *            is the existing UUID to be applied to the new Tree.
 	 */
@@ -114,7 +114,7 @@ public class Base extends StationaryElement implements Ownable, Damageable
 
 	/**
 	 * Checks whether or not this base is currently charging a tank.
-	 * 
+	 *
 	 * @return the current charging status of this base.
 	 */
 	public boolean isCharging()
@@ -124,7 +124,7 @@ public class Base extends StationaryElement implements Ownable, Damageable
 
 	/**
 	 * Sets the charging state of this base.
-	 * 
+	 *
 	 * @param charge
 	 *            represents whether or not this base should be in a charging state. False
 	 *            = not charging!
@@ -136,33 +136,33 @@ public class Base extends StationaryElement implements Ownable, Damageable
 
 	/**
 	 * Returns the current health of the base
-	 * 
+	 *
 	 * @return current hit point count
 	 */
 	@Override
-	public int getHitPoints() 
+	public float getHitPoints()
 	{
 		return hitPoints;
 	}
 
 	/**
-	 * Method that returns the maximum number of hit points the entity can have. 
+	 * Method that returns the maximum number of hit points the entity can have.
 	 * @return - Max Hit points for the entity
 	 */
 	@Override
-	public int getMaxHitPoints() 
+	public int getMaxHitPoints()
 	{
 		return MAX_HIT_POINTS;
 	}
-	
+
 	/**
 	 * Changes the hit point count after taking damage
-	 * 
+	 *
 	 * @param damagePoints
 	 *            how much damage the base has taken
 	 */
 	@Override
-	public void takeHit(int damagePoints) 
+	public void takeHit(int damagePoints)
 	{
 		hitPoints -= Math.abs(damagePoints);
 		if (this.hitPoints <= 0)
@@ -180,11 +180,11 @@ public class Base extends StationaryElement implements Ownable, Damageable
 
 	/**
 	 * Increments the base's health by a given amount
-	 * 
+	 *
 	 * @param healPoints - how many points the base is given
 	 */
 	@Override
-	public void heal(int healPoints) 
+	public void heal(float healPoints)
 	{
 		if (hitPoints + Math.abs(healPoints) < MAX_HIT_POINTS)
 		{
@@ -194,24 +194,24 @@ public class Base extends StationaryElement implements Ownable, Damageable
 		else
 		{
 			hitPoints = MAX_HIT_POINTS;
-		}		
+		}
 	}
-	
+
 	/**
 	 * The current amount of ammo at the base
 	 * @return the current amount of ammo
 	 */
-	public int getAmmoCount() 
+	public int getAmmoCount()
 	{
 		return ammoCount;
 	}
-	
+
 
 	/**
 	 * The maximum amount of ammo a base can have
 	 * @return the maximum amount of ammo storage at a base
 	 */
-	public static int getMaxAmmoCount() 
+	public static int getMaxAmmoCount()
 	{
 		return MAX_AMMO_COUNT;
 	}
@@ -225,12 +225,12 @@ public class Base extends StationaryElement implements Ownable, Damageable
 		{
 			ammoCount += AMMO_REPLENISH_RATE;
 		}
-		else 
+		else
 		{
 			ammoCount = MAX_AMMO_COUNT;
 		}
 	}
-	
+
 	/**
 	 * Method that deducts ammo from supply to give to tank
 	 * @return - amount of ammo capable of being supplied at request
@@ -243,31 +243,31 @@ public class Base extends StationaryElement implements Ownable, Damageable
 			ammoCount = 0;
 			return ammoGiven;
 		}
-		else 
+		else
 		{
 			ammoCount -= AMMO_REPLENISH_RATE;
 			return AMMO_REPLENISH_RATE;
 		}
 	}
-	
+
 	/**
 	 * The current number of mines at a base
 	 * @return the current number of mines
 	 */
-	public int getMineCount() 
+	public int getMineCount()
 	{
 		return mineCount;
 	}
-	
+
 	/**
 	 * The maximum number of mines a base can have
 	 * @return the maximum storage for mines at a base
 	 */
-	public static int getMaxMineCount() 
+	public static int getMaxMineCount()
 	{
 		return MAX_MINE_COUNT;
 	}
-	
+
 	/**
 	 * Replenishes the mines for the base
 	 */
@@ -282,7 +282,7 @@ public class Base extends StationaryElement implements Ownable, Damageable
 			mineCount = MAX_MINE_COUNT;
 		}
 	}
-	
+
 	/**
 	 * Method that deducts mines from supply to give to tank
 	 * @return - amount of mines capable of being supplied at request
@@ -295,23 +295,23 @@ public class Base extends StationaryElement implements Ownable, Damageable
 			mineCount = 0;
 			return minesGiven;
 		}
-		else 
+		else
 		{
 			mineCount -= MINE_REPLENISH_RATE;
 			return MINE_REPLENISH_RATE;
 		}
 	}
 
-	
+
 	/**
 	 * Donates hit points to an object to heal
 	 * @return the amount of hit points to give
 	 */
-	public int giveHitPoints()
+	public float giveHitPoints()
 	{
 		if(hitPoints - HIT_POINTS_REPLENISH_RATE < 0)
 		{
-			int hitPointsGiven = hitPoints;
+			float hitPointsGiven = hitPoints;
 			hitPoints = 0;
 			return hitPointsGiven;
 		}
@@ -321,15 +321,15 @@ public class Base extends StationaryElement implements Ownable, Damageable
 			return HIT_POINTS_REPLENISH_RATE;
 		}
 	}
-	
+
 	@Override
-	public UUID getOwnerUID() 
+	public UUID getOwnerUID()
 	{
 		return this.ownerUID;
 	}
 
 	@Override
-	public void setOwnerUID(UUID ownerUID) 
+	public void setOwnerUID(UUID ownerUID)
 	{
 		this.ownerUID = ownerUID;
 	}
